feat: open upload dialog from nav
